project, service: move version bundle content to pkg/project
Towards https://github.com/giantswarm/giantswarm/issues/7722
