+++ service/controller/clusterapi/v22/cluster_resource_set.go
@@ -21,6 +21,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/cluster-api/pkg/client/clientset_generated/clientset"
 
+	"github.com/giantswarm/cluster-operator/pkg/project"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/controllercontext"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/key"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/app"
@@ -400,7 +401,7 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 			return false
 		}
 
-		if key.OperatorVersion(&cr) == VersionBundle(config.Provider).Version {
+		if key.OperatorVersion(&cr) == project.VersionBundle(config.Provider).Version {
 			return true
 		}
