+++ service/version_bundles.go
@@ -3,13 +3,13 @@ package service
 import (
 	"github.com/giantswarm/versionbundle"
 
-	clusterapiv22 "github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22"
+	"github.com/giantswarm/cluster-operator/pkg/project"
 )
 
 func NewVersionBundles(p string) []versionbundle.Bundle {
 	var versionBundles []versionbundle.Bundle
 
-	versionBundles = append(versionBundles, clusterapiv22.VersionBundle(p))
+	versionBundles = append(versionBundles, project.VersionBundle(p))
 
 	return versionBundles
 }
